+++ lib/cs/BigBrother.Frontend.Shared/ViewModels/ViewModelBase.cs
@@ -7,7 +7,7 @@ public abstract class ViewModelBase : INotifyPropertyChanged
 {
     public event PropertyChangedEventHandler? PropertyChanged;
 
-    protected void OnPropertyChanged([CallerMemberName] string? propertyName = null)
+    protected virtual void OnPropertyChanged([CallerMemberName] string? propertyName = null)
     {
         PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
     }
@@ -16,6 +16,7 @@ public abstract class ViewModelBase : INotifyPropertyChanged
     {
         if (EqualityComparer<T>.Default.Equals(field, value))
             return false;
+
         field = value;
         OnPropertyChanged(propertyName);
         return true;